+++ corehq/__init__.py
@@ -183,7 +183,7 @@ INDICATOR_ADMIN_INTERFACES = (
 from corehq.apps.announcements.interface import (
     ManageGlobalHQAnnouncementsInterface,
     ManageReportAnnouncementsInterface,
-    AccountingInterface2)
+)
 
 ANNOUNCEMENTS_ADMIN_INTERFACES = (
     (_("Manage Announcements"), (
@@ -192,9 +192,13 @@ ANNOUNCEMENTS_ADMIN_INTERFACES = (
     )),
 )
 
+from corehq.apps.accounting.interface import (
+    AccountingInterface
+)
+
 ACCOUNTING_ADMIN_INTERFACES = (
     (_("Manage Announcements"), (
-        AccountingInterface2,
+        AccountingInterface,
     )),
 )
